исправил float на нумерик,

+++ modules/db/createSQL/init.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS storages (
                                         id SERIAL PRIMARY KEY,
-                                        max_weight FLOAT
+                                        max_weight NUMERIC(15, 3)
 );
 
 
@@ -8,8 +8,8 @@ CREATE TABLE IF NOT EXISTS products (
                                         id SERIAL PRIMARY KEY,
                                         article INTEGER NOT NULL UNIQUE,
                                         product_name VARCHAR(255) NOT NULL,
-    storage_id INTEGER,
-    delivery_date DATE,
-    expire_date DATE,
-    weight FLOAT NOT NULL
+                                        storage_id INTEGER,
+                                        delivery_date DATE,
+                                        expire_date DATE,
+                                        weight FLOAT NOT NULL
     );
